Adjust data imports for refactor
