+++ helpinghands/ai/gpt.py
@@ -1,5 +1,5 @@
 from ..utility.logger import get_logger
-from ..utility.data import backup_df
+from ..data.various import backup_df
 
 # from ..utility.decorator import retry
