+++ helpinghands/audio/music.py
@@ -3,7 +3,7 @@ from scipy.io.wavfile import write
 
 import random, pathlib
 
-from ..utility.data import choose_random_file
+from ..data.various import choose_random_file
 from ..audio.processing import (
     bpm_match_two_files,
     play_sound,
